requirements.txt
bcrypt==3.1.7
bleach==3.1.1
blinker==1.4
cffi==1.14.0
Click==7.0
Flask==1.1.1
Flask-Admin==1.5.5
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Pillow==7.0.0
pycparser==2.19
PyMySQL==0.9.3
six==1.14.0
SQLAlchemy==1.3.13
webencodings==0.5.1
Werkzeug==0.16.1
WTForms==2.2.1
requests==2.23.0
